added methods for update or replace plotly data
update is used for data heavy traces that simply update the x,y,z
and text arrays in the data, but leave all the other objects/arrays
in the data array alone, then call plotly.restyle.

replace removes the previous trace and adds a new one in it's place.
in this method, it is a deep replacement of all arrays and objects
stored in the data that has the same name property