fix bug for importing subsets
The import subsets button was inadvertently checking if the analysis
"run" button was active so it would only work if the user had already
done all the steps necessary to run an analysis. This created a chicken
and egg problem because importing the subsets is one of the necessary
steps to running an analysis. This problem is now fixed.